pyproject.toml
[project]
name = "torcheggroll"
version = "0.1.0"
description = "EggRoll-style Evolution Strategies with low-rank noise for PyTorch"
readme = "README.md"
license = "MIT"
authors = [
    { name = "trulite", email = "satya@bluecopa.com" }
]
requires-python = ">=3.10"
keywords = ["evolution-strategies", "pytorch", "optimization", "machine-learning", "low-rank"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]
examples = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/ai4flab/torcheggroll"
Repository = "https://github.com/ai4flab/torcheggroll"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/torcheggroll"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
